Create helper.py, add state parameter for pollution rental forecast
