+++ app/helper.py
@@ -77,11 +77,35 @@ def pollution(city:str):
     return response
 
 
-@router.get('/pollution')
-async def get_pollution(city:str):
-    """ Input: City, 2 letter abbreviation for state
-    Returns a list containing WalkScore, BusScore, and BikeScore in that order
-    """    
-    response = pollution(city)
+
+def what_message(score):
+    if 90 <= score <= 100:
+        return "daily errands do not require a car"
+    elif 70 <= score <= 89:
+        return "most errands can be accomplished on foot"
+    elif 50 <= score <= 69:
+        return "some errands can be accomplished on foot"
+    elif 25 <= score <= 49:
+        return "most errands require a car"
+    else:
+        return " almost all errands require a car"
+
+
+def just_walk_score(address: str = "7 New Port Beach, Louisiana",
+    lat: float = 39.5984,
+    lon: float = -74.2151
+    ):
+
+    walk_api = WalkScoreAPI(api_key= os.getenv('walk_api'))
+
+    result = walk_api.get_score(longitude = lon, 
+            latitude = lat,
+            address = address)
     
+    message = what_message(result.walk_score)
+
+    response = {"walk_score": result.walk_score, 
+                "walk_message":message, 
+                "transit_score": result.transit_score, 
+                "bike_score": result.bike_score}
     return response
