+++ app/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 import uvicorn
 
-from app import db, ml, pollution, external, forecast
+from app import db, ml, external
 
 description = """
 Edit your app's title and description. See [https://fastapi.tiangolo.com/tutorial/metadata/](https://fastapi.tiangolo.com/tutorial/metadata/)
@@ -23,9 +23,8 @@ app = FastAPI(
 
 app.include_router(db.router, tags=['Database'])
 app.include_router(ml.router, tags=['Machine Learning'])
-app.include_router(pollution.router, tags=['Pollution'])
 app.include_router(external.router, tags=['External Resources'])
-app.include_router(forecast.router, tags=['Rent Forecast'])
+
 
 app.add_middleware(
     CORSMiddleware,
